eval_metrics: cast bf16 → float32 in plot_causal_importance_vals too
Same numpy-doesn't-support-BFloat16 bug as 2dd404f8, just one function
up. Used by permuted_ci_plots and uv_plots (both slow=True). Neither is
in the canonical eval set today, so this is preventive — would have
crashed the same way if anyone enabled them.

+++ param_decomp_lab/eval_metrics/plotting.py
@@ -61,7 +61,7 @@ def _plot_causal_importances_figure(
     images = []
     for j, (mask_name, mask) in enumerate(ci_vals.items()):
         # mask has shape (batch, C) or (batch, pos, C)
-        mask_data = mask.detach().cpu().numpy()
+        mask_data = mask.detach().float().cpu().numpy()
         if has_pos_dim:
             assert mask_data.ndim == 3
             mask_data = mask_data[:, 0, :]
